+++ __docs__/wonder-blocks-timing/use-timeout.stories.tsx
@@ -3,6 +3,10 @@ import * as React from "react";
 import {View} from "@khanacademy/wonder-blocks-core";
 import Button from "@khanacademy/wonder-blocks-button";
 
+import {
+    ClearPolicy,
+    SchedulePolicy,
+} from "../../packages/wonder-blocks-timing/src/util/policies";
 import {useTimeout} from "@khanacademy/wonder-blocks-timing";
 
 export default {
@@ -15,20 +19,51 @@ export default {
     },
 };
 
-export const BasicUsage = () => {
+export const Immediately = () => {
     const [callCount, setCallCount] = React.useState(0);
-    const [active, setActive] = React.useState(false);
     const callback = React.useCallback(() => {
         setCallCount((callCount) => callCount + 1);
     }, []);
-    useTimeout(callback, 1000, active);
+    const {isSet, set, clear} = useTimeout(callback, 1000);
     return (
         <View>
+            <View>isSet = {isSet.toString()}</View>
             <View>callCount = {callCount}</View>
-            <View>active = {active.toString()}</View>
-            <Button onClick={() => setActive(!active)} style={{width: 200}}>
-                Toggle active
-            </Button>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={set}>Set timeout</Button>
+                <Button
+                    onClick={() => {
+                        if (clear) {
+                            clear();
+                        }
+                    }}
+                >
+                    Clear timeout
+                </Button>
+            </View>
+        </View>
+    );
+};
+
+export const OnDemandAndResolveOnClear = () => {
+    const [callCount, setCallCount] = React.useState(0);
+    const callback = React.useCallback(() => {
+        // eslint-disable-next-line no-console
+        console.log("action called");
+        setCallCount((callCount) => callCount + 1);
+    }, []);
+    const {isSet, set, clear} = useTimeout(callback, 1000, {
+        clearPolicy: ClearPolicy.Resolve,
+        schedulePolicy: SchedulePolicy.OnDemand,
+    });
+    return (
+        <View>
+            <View>isSet = {isSet.toString()}</View>
+            <View>callCount = {callCount}</View>
+            <View style={{flexDirection: "row"}}>
+                <Button onClick={() => set()}>Set timeout</Button>
+                <Button onClick={() => clear()}>Clear timeout</Button>
+            </View>
         </View>
     );
 };