+++ packages/wonder-blocks-accordion/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-accordion
 
+## 1.3.2
+
+### Patch Changes
+
+-   Updated dependencies [5dfac06e]
+    -   @khanacademy/wonder-blocks-core@6.4.1
+    -   @khanacademy/wonder-blocks-clickable@4.2.2
+    -   @khanacademy/wonder-blocks-icon@4.1.1
+    -   @khanacademy/wonder-blocks-typography@2.1.12
+
 ## 1.3.1
 
 ### Patch Changes