+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @khanacademy/wonder-blocks-pill
 
+## 2.4.1
+
+### Patch Changes
+
+-   Updated dependencies [5dfac06e]
+    -   @khanacademy/wonder-blocks-core@6.4.1
+    -   @khanacademy/wonder-blocks-clickable@4.2.2
+    -   @khanacademy/wonder-blocks-link@6.1.2
+    -   @khanacademy/wonder-blocks-typography@2.1.12
+
+## 2.4.0
+
+### Minor Changes
+
+-   bf1e3d87: Add tabindex to pill props
+
+## 2.3.0
+
+### Minor Changes
+
+-   e8d25c90: Aria props are now properly applied to pill component.
+
 ## 2.2.1
 
 ### Patch Changes